refactor(docs): extract _assert_dataset_table_inlines_datatables into a Sphinx extension
Moves _assert_dataset_table_inlines_datatables (event: builder-inited) from
conf.py into docs/source/_extensions/assert_dataset_table.py. Sanity-checks
that prepare_summary_tables.py inlined the DataTables/jQuery CDN scripts
into the dataset_summary table, failing fast otherwise.
Helpers moved alongside: none (self-contained). conf.py shrinks from 4531
to 4485 LOC.

Part of C1 in docs_pipeline_architecture_review.md
